+++ text2sparql_client/context.py
@@ -1,8 +1,8 @@
 """Context object for the CLI"""
 
-from datetime import UTC, datetime
+import sys
 
-import click
+from loguru import logger
 
 
 class ApplicationContext:
@@ -10,14 +10,8 @@ class ApplicationContext:
 
     def __init__(self, debug: bool):
         self.debug = debug
-
-    def echo_debug(self, message: str) -> None:
-        """Output a debug message if --debug is enabled."""
+        logger.remove()
         if self.debug:
-            now = datetime.now(tz=UTC)
-            click.secho(f"[{now!s}] {message}", err=True, dim=True)
-
-    @staticmethod
-    def echo_info(message: str, nl: bool = True, fg: str = "") -> None:
-        """Output an info message"""
-        click.secho(message, nl=nl, fg=fg, err=True)
+            logger.add(sys.stderr, level="DEBUG")
+        else:
+            logger.add(sys.stderr, level="INFO")