+++ text2sparql_client/request.py
@@ -2,40 +2,44 @@
 
 from datetime import UTC, datetime
 
-from pydantic import BaseModel
 from requests import get
 
-from text2sparql_client.sqlite import Database
-
-
-class ResponseMessage(BaseModel):
-    """Endpoint Response (pydantic model)"""
-
-    dataset: str
-    question: str
-    query: str
-    endpoint: str | None = None
+from text2sparql_client.database import Database
+from text2sparql_client.models.response import ResponseMessage
 
 
 def text2sparql(
     endpoint: str, dataset: str, question: str, timeout: int, database: Database
 ) -> ResponseMessage:
     """Text to SPARQL Request."""
-    response = get(
-        url=endpoint,
-        params={
-            "dataset": dataset,
-            "question": question,
-        },
-        timeout=timeout,
-    )
-    database.add_response(
-        time=str(datetime.now(tz=UTC)),
+    timestamp = str(datetime.now(tz=UTC))
+    database.register_question(
+        time=timestamp,
         endpoint=endpoint,
         dataset=dataset,
         question=question,
-        response=response,
     )
+    try:
+        response = get(
+            url=endpoint,
+            params={
+                "dataset": dataset,
+                "question": question,
+            },
+            timeout=timeout,
+        )
+        database.add_response(
+            time=timestamp,
+            endpoint=endpoint,
+            dataset=dataset,
+            question=question,
+            response=response,
+        )
+    except Exception as error:
+        database.add_exception(
+            time=timestamp, endpoint=endpoint, dataset=dataset, question=question, exception=error
+        )
+        raise
     response_message = ResponseMessage(**response.json())
     response_message.endpoint = endpoint
     return response_message